Moved carbon kernel utils to a separate component
