+++ core/pom.xml
@@ -37,6 +37,11 @@
 
     <dependencies>
         <dependency>
+            <groupId>org.wso2.carbon</groupId>
+            <artifactId>org.wso2.carbon.utils</artifactId>
+            <optional>true</optional>
+        </dependency>
+        <dependency>
             <groupId>org.wso2.eclipse.osgi</groupId>
             <artifactId>org.eclipse.osgi</artifactId>
             <optional>true</optional>
@@ -152,6 +157,7 @@
             org.wso2.carbon.kernel.*; version="${carbon.kernel.package.export.version}",
         </export.package>
         <import.package>
+            org.wso2.carbon.kerenl.utils.*;version="${org.eclipse.osgi.util.package.import.version.range}",
             org.eclipse.osgi.util;version="${org.eclipse.osgi.util.package.import.version.range}",
             org.slf4j.*;version="${slf4j.logging.package.import.version.range}",
             org.osgi.framework.*;version="${osgi.framework.package.import.version.range}",
