+++ core/src/main/java/org/wso2/carbon/kernel/configprovider/utils/ConfigurationUtils.java
@@ -20,7 +20,7 @@ import org.json.JSONObject;
 import org.json.XML;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.wso2.carbon.kernel.Constants;
+import org.wso2.carbon.kernel.utils.Constants;
 import org.yaml.snakeyaml.Yaml;
 
 import java.io.IOException;
@@ -89,7 +89,7 @@ public class ConfigurationUtils {
 
     /**
      * This method reads project properties in resource file,
-     * {@value org.wso2.carbon.kernel.Constants#PROJECT_DEFAULTS_PROPERTY_FILE}
+     * {@value org.wso2.carbon.kernel.utils.Constants#PROJECT_DEFAULTS_PROPERTY_FILE}
      *
      * @return project properties
      */
