+++ core/src/main/java/org/wso2/carbon/kernel/internal/CarbonStartupHandler.java
@@ -18,6 +18,7 @@ package org.wso2.carbon.kernel.internal;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.kernel.utils.CarbonServerInfo;
+import org.wso2.carbon.kernel.utils.Constants;
 
 import java.text.DecimalFormat;
 
@@ -39,7 +40,7 @@ public class CarbonStartupHandler {
      * @param serverName Server name to be in the log
      */
     public static void logServerStartupTime(String serverName) {
-        double startTime = Long.parseLong(System.getProperty(org.wso2.carbon.kernel.Constants.START_TIME));
+        double startTime = Long.parseLong(System.getProperty(Constants.START_TIME));
         double startupTime = (System.currentTimeMillis() - startTime) / 1000;
 
         DecimalFormat decimalFormatter = new DecimalFormat("#,##0.000");
