+++ core/src/main/java/org/wso2/carbon/kernel/jmx/security/CarbonJMXAuthenticator.java
@@ -15,7 +15,7 @@
  */
 package org.wso2.carbon.kernel.jmx.security;
 
-import org.wso2.carbon.kernel.Constants;
+import org.wso2.carbon.kernel.utils.Constants;
 
 import java.util.Collections;
 import javax.management.remote.JMXAuthenticator;
