+++ core/src/test/java/org/wso2/carbon/kernel/configprovider/ConfigProviderImplTest.java
@@ -22,11 +22,11 @@ import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
-import org.wso2.carbon.kernel.Constants;
 import org.wso2.carbon.kernel.internal.configprovider.ConfigProviderDataHolder;
 import org.wso2.carbon.kernel.internal.configprovider.ConfigProviderImpl;
 import org.wso2.carbon.kernel.securevault.SecureVault;
 import org.wso2.carbon.kernel.securevault.exception.SecureVaultException;
+import org.wso2.carbon.kernel.utils.Constants;
 import org.wso2.carbon.kernel.utils.EnvironmentUtils;
 import org.yaml.snakeyaml.Yaml;
