+++ core/src/test/java/org/wso2/carbon/kernel/context/CarbonContextTest.java
@@ -18,10 +18,10 @@ package org.wso2.carbon.kernel.context;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.wso2.carbon.kernel.CarbonRuntime;
-import org.wso2.carbon.kernel.Constants;
 import org.wso2.carbon.kernel.configprovider.ConfigProvider;
 import org.wso2.carbon.kernel.internal.context.CarbonConfigProviderImpl;
 import org.wso2.carbon.kernel.internal.context.CarbonRuntimeFactory;
+import org.wso2.carbon.kernel.utils.Constants;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
