+++ core/src/test/java/org/wso2/carbon/kernel/securevault/DefaultSecretRepositoryTest.java
@@ -20,11 +20,11 @@ import org.easymock.EasyMock;
 import org.testng.Assert;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
-import org.wso2.carbon.kernel.Constants;
 import org.wso2.carbon.kernel.securevault.config.model.SecretRepositoryConfiguration;
 import org.wso2.carbon.kernel.securevault.exception.SecureVaultException;
 import org.wso2.carbon.kernel.securevault.repository.DefaultSecretRepository;
 import org.wso2.carbon.kernel.securevault.utils.DefaultHardCodedMasterKeyReader;
+import org.wso2.carbon.kernel.utils.Constants;
 
 import java.io.File;
 import java.io.FileOutputStream;
