+++ core/src/test/java/org/wso2/carbon/kernel/utils/MBeanRegistratorTest.java
@@ -20,7 +20,6 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import org.wso2.carbon.kernel.Constants;
 import org.wso2.carbon.kernel.transports.TransportManager;
 
 import javax.management.InstanceNotFoundException;
