+++ core/src/test/resources/testng.xml
@@ -36,13 +36,10 @@ limitations under the License.
 
             <class name="org.wso2.carbon.kernel.runtime.RuntimeManagerTest" />
 
-            <class name="org.wso2.carbon.kernel.utils.FileUtilsTest" />
             <class name="org.wso2.carbon.kernel.internal.DataHolderTest" />
 
             <class name="org.wso2.carbon.kernel.internal.context.DefaultCarbonRuntimeTest" />
             <class name="org.wso2.carbon.kernel.internal.context.CarbonRuntimeFactoryTest" />
-            <class name="org.wso2.carbon.kernel.utils.UtilsTest"/>
-            <class name="org.wso2.carbon.kernel.utils.manifest.ManifestElementTest"/>
             <class name="org.wso2.carbon.kernel.utils.MBeanManagementFactoryTest"/>
             <class name="org.wso2.carbon.kernel.utils.MBeanRegistratorTest"/>
