+++ pom.xml
@@ -43,6 +43,7 @@
             <modules>
                 <module>parent</module>
                 <module>archetypes</module>
+                <module>utils</module>
                 <module>launcher</module>
                 <module>core</module>
                 <module>extensions</module>
