+++ tests/osgi-tests/pom.xml
@@ -201,6 +201,11 @@
             <artifactId>org.wso2.carbon.sample.runtime.service</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon</groupId>
+            <artifactId>org.wso2.carbon.utils</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
     <profiles>
         <profile>
