+++ tests/osgi-tests/src/test/java/org/wso2/carbon/osgi/context/CarbonContextOSGiTest.java
@@ -29,10 +29,10 @@ import org.testng.annotations.Test;
 import org.wso2.carbon.container.CarbonContainerFactory;
 import org.wso2.carbon.container.options.CopyFileOption;
 import org.wso2.carbon.context.test.CarbonContextInvoker;
-import org.wso2.carbon.kernel.Constants;
 import org.wso2.carbon.kernel.context.CarbonContext;
 import org.wso2.carbon.kernel.context.PrivilegedCarbonContext;
 import org.wso2.carbon.kernel.utils.CarbonServerInfo;
+import org.wso2.carbon.kernel.utils.Constants;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
