+++ utils/src/main/java/org/wso2/carbon/kernel/utils/CarbonServerInfo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 WSO2, Inc. http://www.wso2.org
+ * Copyright 2017 WSO2, Inc. http://www.wso2.org
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
