+++ utils/src/main/java/org/wso2/carbon/kernel/utils/Constants.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.wso2.carbon.kernel;
+package org.wso2.carbon.kernel.utils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,7 +21,7 @@ import org.slf4j.LoggerFactory;
 /**
  * Carbon Constants.
  *
- * @since 5.0.0
+ * @since 5.2.0
  */
 public final class Constants {
