+++ utils/src/main/java/org/wso2/carbon/kernel/utils/manifest/ManifestElement.java
