+++ utils/src/test/resources/sample.txt
@@ -1,0 +1,1 @@
+Test resources for org.wso2.carbon.utils.FileUtils.java
